Added: user interceptor and decorator to authorize the req

+++ src/home/home.controller.ts
@@ -8,10 +8,12 @@ import {
   Post,
   Put,
   Query,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { HomeService } from './home.service';
 import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dtos/home.dto';
 import { PropertyType } from '@prisma/client';
+import { User } from 'src/user/decorators/user.decorator';
 
 @Controller('home')
 export class HomeController {
@@ -45,20 +47,29 @@ export class HomeController {
   }
 
   @Post()
-  createHome(@Body() body: CreateHomeDto) {
-    return this.homeService.createHome(body);
+  createHome(@Body() body: CreateHomeDto, @User() user: User) {
+    return this.homeService.createHome(body, user.id);
   }
 
   @Put(':id')
-  updateHome(
+  async updateHome(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: UpdateHomeDto,
+    @User() user: User,
   ) {
+    const realtorId = await this.homeService.getRealorByHomeId(id);
+    if (realtorId.id !== user.id) {
+      throw new UnauthorizedException('Unauthorized');
+    }
     return this.homeService.updateHome(id, body);
   }
 
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(@Param('id', ParseIntPipe) id: number, @User() user: User) {
+    const realtorId = await this.homeService.getRealorByHomeId(id);
+    if (realtorId.id !== user.id) {
+      throw new UnauthorizedException('Unauthorized');
+    }
     return this.homeService.deleteHome(id);
   }
 }
